ChronoNet/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.10
django-crispy-forms==1.13.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
python-dotenv==0.19.2
Pillow
django-admin-honeypot
djangorestframework
django-markdownfield>=0.10.0
psycopg2-binary